Add orderfrom, orderto, and orderstatus commands as well as GoogleMapAPI class to control googleMap API usage.
Add orderfrom, orderto, and orderstatus commands as well as GoogleMapAPI class. In addition to that, add a method in DatabaseCon to be used by orderfrom command and some commands in case the user is not in group, is not logged in, or the user mispell/type a command that does not exist.